govaudit_ai_pro3/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.11
python-multipart==0.0.6
aiofiles==23.1.0
streamlit==1.40.2
requests==2.32.3
pytesseract==0.3.13
pillow==10.4.0
pdf2image==1.17.0
